+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 
 ### Changed
 
+- Add "page-break-before: avoid" to short final list items
+
 ### Fixed
 
 - Show the right heading level for callout boxes with headings